requirements/base.txt
python-slugify==8.0.4  # https://github.com/un33k/python-slugify
Pillow==10.3.0  # https://github.com/python-pillow/Pillow
argon2-cffi==23.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.6.0  # https://github.com/evansd/whitenoise
redis==5.0.4  # https://github.com/redis/redis-py
hiredis==2.3.2  # https://github.com/redis/hiredis-py

# Django
# ------------------------------------------------------------------------------
django==4.2.13  # pyup: < 5.0  # https://www.djangoproject.com/
django-environ==0.11.2  # https://github.com/joke2k/django-environ
django-model-utils==4.5.1  # https://github.com/jazzband/django-model-utils
django-allauth[mfa]==0.63.1  # https://github.com/pennersr/django-allauth
django-crispy-forms==2.1  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==2024.2  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==5.4.0  # https://github.com/jazzband/django-redis
factory-boy==3.3.0
faker==25.0.0
django-filter==24.2
djangorestframework==3.15.1
django-import-export==4.0.0
django-widget-tweaks==1.5.0
django-after-response==0.2.2
django-auditlog==3.0.0
django-silk==5.1.0
factory-boy==3.3.0  # https://github.com/FactoryBoy/factory_boy
django-debug-toolbar==4.3.0  # https://github.com/jazzband/django-debug-toolbar
django-extensions==3.2.3
django-cors-headers==4.3.1